api/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.1
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.9
cryptography==43.0.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.4.1
more-itertools==10.5.0
numpy==2.3.0
pandas==2.3.0
paramiko==3.5.0
pendulum==3.0.0
pgcli==4.1.0
pgspecial==2.1.2
pillow==11.2.1
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-c==3.2.3
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
SecretStorage==3.3.3
setproctitle==1.3.3
six==1.16.0
sqlparse==0.5.1
sshtunnel==0.4.0
tabulate==0.9.0
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
